End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -60,13 +61,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
 
-        //TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
         
         _context.Auctions.Add(auction);
 
@@ -82,15 +83,16 @@ public class AuctionsController : ControllerBase
             new {auction.Id}, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
-
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions.Include(x => x.Item)
             .FirstOrDefaultAsync(x => x.Id == id);
 
-        //TODO: check Seller == username
-        
+        //check Seller == username, while updating the auction
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color; 
@@ -106,6 +108,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");     
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -113,8 +116,9 @@ public class AuctionsController : ControllerBase
 
         if(auction == null) return NotFound();
 
-        //TODO: check Seller == username
-        
+        //check Seller == username, while deleting auction
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
         _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>(new {id = auction.Id.ToString()});
